+++ ohsome_quality_api/indicators/building_comparison/indicator.py
@@ -233,6 +233,21 @@ class BuildingComparison(BaseIndicator):
         }
         fig.update_layout(**layout)
 
+        fig.add_layout_image(
+            dict(
+                source="https://raw.githubusercontent.com/GIScience/ohsome-quality-api/refs/heads/figure_background_logo/ohsome_quality_api/indicators/heigit_logo.png",
+                xref="paper",
+                yref="paper",
+                x=0.8,
+                y=-0.15,
+                sizex=0.2,
+                sizey=0.2,
+                sizing="contain",
+                opacity=0.3,
+                layer="above",
+            )
+        )
+
         raw = fig.to_dict()
         raw["layout"].pop("template")  # remove boilerplate
         self.result.figure = raw