+++ ohsome_quality_api/indicators/land_cover_completeness/indicator.py
@@ -107,6 +107,21 @@ class LandCoverCompleteness(BaseIndicator):
         fig.update_xaxes(visible=False)
         fig.update_yaxes(visible=False)
 
+        fig.add_layout_image(
+            dict(
+                source="https://raw.githubusercontent.com/GIScience/ohsome-quality-api/refs/heads/figure_background_logo/ohsome_quality_api/indicators/heigit_logo.png",
+                xref="paper",
+                yref="paper",
+                x=0.41,
+                y=0.55,
+                sizex=0.2,
+                sizey=0.2,
+                sizing="contain",
+                opacity=0.3,
+                layer="above",
+            )
+        )
+
         raw = fig.to_dict()
         raw["layout"].pop("template")  # remove boilerplate
         self.result.figure = raw