+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### New Features
 
 - new topics about cycleways and bridges ([#928])
-- land-cover-thematic-accuracy: add warning if aoi is not completly covered by the comparison dat ([#930])
+- land-cover-thematic-accuracy: add warning if AOI is not completely covered by the comparison data ([#930])
 
 [#928]: https://github.com/GIScience/ohsome-quality-api/pull/928
 [#930]: https://github.com/GIScience/ohsome-quality-api/pull/930